requirements.txt
# Philips Hue API
phue==1.1

# Data analysis
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Scheduling
schedule==1.2.1
apscheduler==3.10.4

# Database
sqlalchemy==2.0.23

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging
loguru==0.7.2

# HTTP requests (backup for Hue API)
requests==2.31.0
